refactor(sqlite): separate quarantine errors from migrations
