+++ src/state/openclaw-quarantine-store.ts
@@ -12,6 +12,7 @@ import {
   type SqliteFileGeneration,
 } from "../infra/sqlite-file-generation.js";
 import { VERSION } from "../version.js";
+import { OPENCLAW_DATABASE_SCHEMA_DOCS_URL } from "./openclaw-state-db-contract.js";
 import { resolveOpenClawStateSqliteDir } from "./openclaw-state-db.paths.js";
 
 const OPENCLAW_QUARANTINE_SCHEMA_VERSION = 2;
@@ -26,6 +27,21 @@ type OpenClawDatabaseQuarantine = {
   quarantinedAt: number;
   reason: string;
 };
+
+// Read admission needs this error without importing schema migrations.
+export function createOpenClawDatabaseVerificationError(
+  kind: "agent" | "state",
+  pathname: string,
+  storedError: string | null,
+): Error {
+  // Doctor's clearing hooks run after a full integrity assertion, so a still-
+  // corrupt file cannot be cleared directly: the file must be healthy first.
+  const error = new Error(
+    `OpenClaw ${kind} database ${pathname} is quarantined after integrity verification failed: ${storedError ?? "unknown integrity error"}. Restore the database from a backup or repair it, then run openclaw doctor --fix to clear the quarantine. See ${OPENCLAW_DATABASE_SCHEMA_DOCS_URL}.`,
+  );
+  error.name = "SqliteIntegrityError";
+  return error;
+}
 
 function resolveQuarantineStorePath(env: NodeJS.ProcessEnv): string {
   return path.join(resolveOpenClawStateSqliteDir(env), "openclaw-quarantine.sqlite");
